+++ apps/demo/src/app/modules/doc/modules/interaction/component/interaction.popover.component.ts
@@ -1,4 +1,4 @@
-import { Component, ElementRef } from '@angular/core';
+import { Component, ElementRef, OnInit } from '@angular/core';
 import { DbxPopoverKey, AbstractPopoverDirective, DbxPopoverService } from '@dereekb/dbx-web';
 import { NgPopoverRef } from 'ng-overlay-container';
 
@@ -23,16 +23,29 @@ export interface DocInteractionPopoverConfig {
       </dbx-popover-controls>
       <dbx-popover-scroll-content>
         <dbx-interaction-example-popover-content (return)="returnAndClose($event)" (close)="close()"></dbx-interaction-example-popover-content>
+        <p class="dbx-hint">Configured to return "{{ onCloseValue }}" when closed via a backdrop click.</p>
       </dbx-popover-scroll-content>
     </dbx-popover-content>
   `
 })
-export class DocInteractionExamplePopoverComponent extends AbstractPopoverDirective<number> {
+export class DocInteractionExamplePopoverComponent extends AbstractPopoverDirective<number> implements OnInit {
+  readonly onCloseValue = 12345;
+
   static openPopover(popoverService: DbxPopoverService, { origin }: DocInteractionPopoverConfig, popoverKey?: DbxPopoverKey): NgPopoverRef<any, number> {
     return popoverService.open({
       key: popoverKey ?? DEFAULT_INTERACTION_POPOVER_COMPOSER_POPOVER_KEY,
       origin,
       componentClass: DocInteractionExamplePopoverComponent
     });
+  }
+
+  ngOnInit(): void {
+    this.popover.getClosingValueFn = (_, type) => {
+      if (type === 'backdropClick') {
+        return this.onCloseValue;
+      } else {
+        return undefined;
+      }
+    };
   }
 }